requirements.txt
python-can==3.3.4